+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 1.6 (unreleased)
 ^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Display results as votes or percentage (closes `#4`_).
+  [marcosfromero]
 
 
 1.5 (2013-12-17)
@@ -170,6 +171,7 @@ There's a frood who really knows where his towel is.
 
 - Initial release.
 
+.. _`#4`: https://github.com/collective/collective.polls/issues/4
 .. _`#35`: https://github.com/collective/collective.polls/issues/35
 .. _`#51`: https://github.com/collective/collective.polls/issues/51
 .. _`#53`: https://github.com/collective/collective.polls/issues/53